Adding markers
I added a button and text field that allows you to add markers with the
given title, as well as a class we can use that would store all the
relevant information for each event. The markers are clickable but they
don't do anything yet; ideally they'd go to a new activity that gave you
more editing options.