Improved upon details activity
Creating an event now takes you to an activity that asks for all
details. I changed the date object for eventMarker to 2 Strings to
simplify the user input process. After the user enters information, if
they press "Create", then all of the fields are added to the temp
object, which we then need to send to the server. If the user presses
cancel, they are sent to the map activity and the marker for that event
is removed.