Deletion of tasks
Tasks can now be viewed after they are created by pressing the marker,
and then pressing the infowindow that appears above the marker. From
this activity, the user can view the task details. The delete button
removes the task from the array. The edit button still needs to be
implemented.