Event Details Layout And Icon
-Made a simple icon in case we run out of time for stuff like that
-Made the event Details layout be nest in a linear layout in order to
prevent overlapping issues. I've also made them scrollable incase the
description and event title go on for a while. Also I got right of the
title bar, turns out it was really easy.
- Removed padding on google map to make it look slightly better.

Apparently all the files were duplicated as well so I got rid of those.
I'll have all of tomorrow to work on this as well.